Refactor: Add fallback for event where web workers aren't avaiable
