+++ static/js/navigation/careers/Worker_populate-careers.js
@@ -6,11 +6,7 @@ self.onmessage = function (e) {
   fetch("/careers/roles.json")
     .then((response) => response.json())
     .then((departments) => {
-      const processedData = departments.map((dpt) => ({
-        slug: dpt.slug,
-        count: dpt.count,
-      }));
-      self.postMessage(processedData);
+      self.postMessage(departments);
     })
     .catch((error) => {
       console.error(`Unable to load careers navigation roles: ${error}`);
